- scripting: Add system for registring typetags from plugins, rebind LibFinder
> There is new call in ScriptingManager which makes it possible to register
  dynamic typetags if you have classes you have to bind and unbind to
  squirrel. Ideally this is done at plugin attach/release and it should be
  a rare thing. If the class is re-registered all old instances would be
  different class, but this is still really bad if it happens.
> Register the lib finder class LibFinder.
> Change the Wiz class to have a dynamic typetag and to be almost properly
  unregistered (if there are alive instances of this class and they are
  used after the class is unregistered there would be crashes).
> Make the UI class/namespace to be private, so other binding code doesn't
  have access to it.

git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@12415 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/include/scripting/bindings/sc_utils.h
@@ -121,7 +121,7 @@ UserDataForType<UserType>* SetupUserPointer(HSQUIRRELVM vm, SQInteger idx)
 {
     SQUserPointer ptr = nullptr;
     if (SQ_FAILED(sq_getinstanceup(vm, idx, &ptr,
-                                   SQUserPointer(TypeInfo<UserType>::typetag))))
+                                   SQUserPointer(uint64_t(TypeInfo<UserType>::typetag)))))
     {
         sq_throwerror(vm, _SC("SetupUserPointer: Invalid type tag!"));
         cbAssert(false);
@@ -599,7 +599,7 @@ inline SQInteger CreateClassDecl(HSQUIRRELVM v, const SQChar *className,
     }
 
     sq_newclass(v, ((baseClassName != nullptr) ? SQTrue : SQFalse));
-    sq_settypetag(v, -1, SQUserPointer(TypeInfo<UserType>::typetag));
+    sq_settypetag(v, -1, SQUserPointer(uint64_t(TypeInfo<UserType>::typetag)));
     // Add some memory to the class. We over-allocate in cases where we store only a pointer.
     sq_setclassudsize(v, -1, sizeof(UserDataForType<UserType>));
